% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_var.R
\name{analyse_var}
\alias{analyse_var}
\title{Analyse variable by ab test groups}
\usage{
analyse_var(df, var_chr, group_variable = "ab_test_group",
  quant_focus = "mid", confidence_interval = 0.9)
}
\arguments{
\item{df}{data frame with both variables of interest}

\item{var_chr}{column name of variable you want to be analysed as string}

\item{group_variable}{variable to group by}

\item{confidence_interval}{probability of confidence interval range}

\item{quant_focut}{("up", "mid", "down") controls which quantiles to focus more.
"up" focuses more on median and upper quantiles, "down" opposite and "mid" is balanced.}
}
\value{
list of two tibbles: first containing quantile information, and second containing mean,
standard deviation, standard error and lower and upped confidence bounds
}
\description{
If you have 2 columns, one representing variable of interest, and another group for that observation
this function analyses distribution of variable in each group in a way easy to compare
}
\examples{
df <- tribble(
  ~x, ~group,
  1, 1,
  3.2, 1,
  2.4, 1,
  3.1, 1,
  5, 2,
  6, 2,
  4.7, 2
)

df \%>\%
  analyse_var("x", "group")

}
\author{
Elio Bartoš
}
