% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bernoulli_beta_model.R
\name{bernoulli_beta_model}
\alias{bernoulli_beta_model}
\title{Run bayesian beta - bernoulli model for estimating proportion of 0-1 variable}
\usage{
bernoulli_beta_model(alpha, beta, success, total, sample_size = 1e+05)
}
\arguments{
\item{alpha}{Parameter for prior distribution representing the number of success}

\item{beta}{Parameter for prior distribution representing the number of fails
Prior used is Beta(alpha + 1, beta + 1)}

\item{success}{Number of success cases in your data}

\item{total}{Total number of cases in your data}

\item{sample_size}{Size of sample from posterior distribution}
}
\value{
Vector of samples from posterior distribution

Posterior distribution is Beta(alpha + 1 + success, beta + 1 + total - success)
}
\description{
Runs a bayesian estimation of proportion using bernoulli distribution as likelihood and beta distribution as conjugate prior.
Posterior distribution is beta distribution.
}
\examples{
post = bernoulli_beta_model(0, 0, 20, 100) # No prior information, pror is uniform
post2 = bernoulli_beta_model(5, 95, 3, 50) # Prior succes rate is around 5\% with estimation strenght as it was estimated on a sample of 100

mean(post)
quantile(post, probs = c(0.05, 0.95)) # 90\% highest density posterior interval

mean(post2)
quantile(post2, probs = c(0.05, 0.95))
}
