% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_chisq.R
\name{check_chisq}
\alias{check_chisq}
\title{Check whether variables can use chisq test.}
\usage{
check_chisq(dat, group, ...)
}
\arguments{
\item{dat}{data.frame}

\item{group}{The group variable with unquoted name}

\item{...}{One or more unquoted expressions separated by commas. If none, use all variables except group.}
}
\value{
A list whose class is \code{check_chisq}.
\describe{
  \item{methods}{
  Named character vector to represent the method for every variable. \code{'chisq'} represents the \code{chisq.test},
  \code{'fisher'} represents the \code{fisher.test}, \code{NA} means that the variable isn't a factor.
  }
  \item{expected}{Named list which contains the expected counts of all variables.}
}
}
\description{
This function use the \code{chisq.test} to calculate the expected frequency, so there often are some warnings
about "Chi-squared approximation may be incorrect".
}
\examples{
# settle data used to test
library(tidyverse)
dat <- mtcars \%>\% mutate(cyl=factor(cyl), vs=factor(vs), am=factor(am), gear=factor(gear), carb=factor(carb))
# test
dat \%>\% check_chisq(cyl)
}
\seealso{
\code{\link{check_NE}}
}
