% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_NE.R
\name{check_NE}
\alias{check_NE}
\title{Check whether Normality and homogeneity of variance are satisfied}
\usage{
check_NE(
  dat,
  group,
  ...,
  .N = TRUE,
  .E = TRUE,
  .Nthre = 0.1,
  .Ethre = 0.1,
  .Ntest_object = "every"
)
}
\arguments{
\item{dat}{data.frame}

\item{group}{The group variable with unquoted name}

\item{...}{One or more unquoted expressions separated by commas. If none, use every variables else.}

\item{.N}{Whether need check normality, default TRUE.}

\item{.E}{Whether need check homogeneity of variance, default TRUE.}

\item{.Nthre}{Significance level for Normalization test, default 0.1.}

\item{.Ethre}{Significance level for variance equalization test, default 0.1.}

\item{.Ntest_object}{character, \code{"all"}, \code{"every"} or \code{"total"}. \code{"total"} means that
normality is checked for a whole variable. \code{"every"} means that normality is checked for every category of
variable. \code{"all"} means above two are both checked. Default "every".}
}
\value{
A list whose class is \code{check_NE}.
\describe{
  \item{methods}{
  Named character vector to represent the methods should used. If the the normality and homogeneity test can both pass,
  the variable uses \code{t.test} or \code{anova}. If the normality test can pass but homogeneity can't, the variable uses
  \code{t.test} (Welch t). If the normality test can't pass, \code{wilcox.test} or \code{kruskal.test} should be used.
  \code{NA} means that the variable isn't a numeric vector. Only the \code{.N=TRUE} and \code{.E=TRUE}, this item will be
  provided.
  }
  \item{logic_N}{Named logical vector to indicate which variable is normality.}
  \item{logic_E}{Named logical vector to indicate which variable is homogeneity of variable.}
  \item{pvalue_N}{Float vector of p value of normality test.}
  \item{pvalue_E}{Float vector of p value of homogeneity test of variable.}
  \item{test_method_N}{Until to now, it's just \code{"shapiro.test"}.}
  \item{test_method_E}{Until to now, it's just \code{"leveneTest"}.}
  \item{test_object_N}{It's the parameter \code{.Ntest_object}.}
}
}
\description{
This function check whether the data satisfies both normality and homogeneity of variance. The \code{shapiro.test}
and \code{car::leveneTest} are used. It's worth noting that we default to check normality for every category and whole variable. Parameter
\code{.Ntest_object} is used to control this. In order to reduce the type II error, \code{.Nthre=0.1} and \code{.Ethre=0.1}
are default.
}
\examples{
check_NE(iris, Species)

# settle data used to test
library(tidyverse)
dat <- mtcars \%>\% mutate(cyl=factor(cyl), vs=factor(vs), am=factor(am), gear=factor(gear), carb=factor(carb))
# test
dat \%>\% check_NE(cyl)
}
\seealso{
\code{\link{check_chisq}}
}
