% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_words_ptbr.R
\name{combine_words_ptbr}
\alias{combine_words_ptbr}
\title{Combine words using ptbr rules (differ from knitr::combine_words() which uses oxford commas)}
\usage{
combine_words_ptbr(words, sep = NULL, last = NULL)
}
\arguments{
\item{words}{a character vector with words to combine}

\item{sep}{a character with the separator of the words. Default is NULL and insert ", "}

\item{last}{a character vector with the last separator of the words. Default is NULL and insert " e "}
}
\value{
a character vector
}
\description{
Combine words using ptbr rules (differ from knitr::combine_words() which uses oxford commas)
}
\examples{
\dontrun{
misc::ipak(c("dplyr", "broom", "glue"))

# showing significant variables
mtcars \%>\%
  select(mpg, cyl, carb, wt) \%>\%
  lm(mpg ~ cyl + carb + wt, data = .) \%>\%
  tidy() \%>\%
  filter(p.value <= 0.05, term != "(Intercept)") \%>\%
  pull(term) \%>\% {
    paste("Variáveis significativas:", combine_words_ptbr(.))
  }

# using in an ordinary text
feira <- c("banana", "maça", "pepino", "ovos")
glue("Por favor, compre: {combine_words_ptbr(feira)}")
}
}
