% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_shap.R
\name{compute_finite_difference_gradients}
\alias{compute_finite_difference_gradients}
\title{Compute finite difference gradients}
\usage{
compute_finite_difference_gradients(f, X, h)
}
\description{
Compute finite difference gradients
}
