% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_shap.R
\name{compute_linear_shap}
\alias{compute_linear_shap}
\title{Compute SHAP values for linear models}
\usage{
compute_linear_shap(model, X, newx, y, account_correlations, n_samples)
}
\description{
Compute SHAP values for linear models
}
