% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_shap.R
\name{correlation_aware_shap}
\alias{correlation_aware_shap}
\title{Correlation-aware SHAP computation using sampling approach}
\usage{
correlation_aware_shap(coefs, newx_centered, X_train, feature_means, n_samples)
}
\description{
Correlation-aware SHAP computation using sampling approach
}
