% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug.R
\name{debug_print}
\alias{debug_print}
\title{Debug print}
\usage{
debug_print(x)
}
\arguments{
\item{x}{An object to be printed}
}
\description{
Debug print
}
\examples{

misc::debug_print(1:10)
misc::debug_print("Hello, world!")

}
