% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc_baseline.R
\name{desc_baseline}
\alias{desc_baseline}
\title{Statistical descripation for baseline variables}
\usage{
desc_baseline(
  dat,
  group,
  ...,
  .num_desc_type = "auto",
  .num_margin = TRUE,
  .fct_margin = "none",
  .fct_prob = "none",
  .check_NE_args = list()
)
}
\arguments{
\item{dat}{data.frame}

\item{group}{The group variable unquoted name}

\item{...}{One or more unquoted expressions separated by commas}

\item{.num_desc_type}{Character, "auto" , "median" or "mean". It's used to control the forms of numeric variable.
"mean" means "mean+-std". "median" means "median (25% quantile, 75% quantile)". "auto" means that forms are
controlled by the \code{check_NE} (Just look the results of normality test).}

\item{.num_margin}{Logical, whether caculate the numeric variables' statistics for all samples. Default TRUE.}

\item{.fct_margin}{Character, "none", "row", "col", "all". Whether margin counts will be calculated. If it's
"row", the counts of each row will be calculated. If "col", the counts of each col will be calculated. if
"all", above two will be both done.}

\item{.fct_prob}{Character, "none", "row", "col", "all". Whether the frenquency will be calculated. "row" means
the frequency will be calculated relative to row margin counts. As for "col", It's col margin counts. If "all",
the total counts will be used.}

\item{.check_NE_args}{If \code{.num_desc_type="auto"}, this is the list of parameters of \code{check_NE} expect \code{dat},
\code{group} and \code{...}. The values are fed as named list. Default is \code{list(.N = TRUE, .E = FALSE, .Nthre = 0.1, .Ethre = 0.1, .Ntest_object = "every")}}
}
\value{
desc_baseline object. As a matter of fact, it's a list.
\describe{
    \item{main}{list, each item is the result for one variable. For numeric variable, the item is a dataframe which
    columns are "mean", "std" for \code{.num_desc_type="mean"} or "median", "quan25" and "quan75" for \code{.num_desc_type="median"}.
    For factor, the item is a dataframe when \code{.fct_prob="none"}. If \code{.fct_prob} is not "none", the item
    will be a list that contains  count and prob.
    }
    \item{desc_methods}{Character vector that indicates the described type.}
    \item{vars, group}{stored original data.}
}
}
\description{
This function can be used to generate the statistical describtion in publishing level. It create a
desc_baseline object, then \code{format} can be used to convert this object to a string dataframe publish.
}
\examples{
desc_baseline(iris, Species) \%>\% format(digit=2)
}
\seealso{
\code{\link{check_NE}} \code{\link{format.desc_baseline}}
}
