% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive_table.R
\name{descriptive_table}
\alias{descriptive_table}
\title{descriptive_table}
\usage{
descriptive_table(data, cols, cor_sig_test = "raw", digit = 3)
}
\arguments{
\item{data}{a dataframe}

\item{cols}{vector or quos(). column(s) that need to be recoded}

\item{cor_sig_test}{adjusted or raw. Default as adjusted. See psych::corr.test to learn more.}

\item{digit}{number of digits}
}
\value{

}
\description{
This function uses the psych::corr.test function to generated the correlation. One potential problem with the current version of the function is that if the correlation between the two item is too high (rounded as 1), then that cell will become blank. This is a very rare problem, and I have not yet figure out how to solve this.
}
\examples{
descriptive_table(iris,cols = tidyr::everything())
}
