% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/destroy_tweet.R
\name{destroy_tweet}
\alias{destroy_tweet}
\title{delete tweets}
\usage{
destroy_tweet(id, token = token, ...)
}
\arguments{
\item{id}{tweet id}

\item{token}{Twitter token created in rtweet}
}
\description{
delete tweets
}
\examples{
\dontrun{
# delete a tweet:
destroy_tweet(id="97690247271840677", token)
}

}
