% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.select_params}
\alias{.select_params}
\title{Select the group variable and other target variables as a vector and data.frame from dat}
\usage{
.select_params(dat, group, ...)
}
\arguments{
\item{dat}{Data}

\item{group}{unquoted variable name for group}

\item{...}{One or more unquoted expressions separated by commas}
}
\value{
list of vector and data.frame
}
\description{
Select the group variable and other target variables as a vector and data.frame from dat
}
\examples{
select_params(iris, Species)
}
