% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_shap.R
\name{extract_coefficients}
\alias{extract_coefficients}
\title{Extract coefficients and intercept from various model types}
\usage{
extract_coefficients(model, model_type)
}
\description{
Extract coefficients and intercept from various model types
}
