% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_shap.R
\name{fit_linear_model}
\alias{fit_linear_model}
\title{Fit linear model based on specified type}
\usage{
fit_linear_model(X, y, model_type, lambda, tau, alpha)
}
\description{
Fit linear model based on specified type
}
