% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_param_dist.R
\name{fit_param_dist}
\alias{fit_param_dist}
\title{Fit multiple parametric distributions, compute KL divergence, simulate best fit}
\usage{
fit_param_dist(vector, num_bins = 30, verbose = TRUE)
}
\arguments{
\item{vector}{Numeric vector of data to fit}

\item{num_bins}{Number of bins for the empirical histogram}

\item{verbose}{Logical indicating whether to print results}
}
\value{
Function to simulate data from the best-fitting distribution
}
\description{
Fit multiple parametric distributions, compute KL divergence, simulate best fit
}
\examples{

set.seed(123)
n <- 1000
vector <- rnorm(n)

start <- proc.time()[3]
simulate_function <- fit_param_dist(vector)
end <- proc.time()[3]
print(paste("Time taken:", end - start))
simulated_data <- simulate_function(n)  # Generate 100 samples from the best-fit distribution
par(mfrow = c(1, 2))
hist(vector, main = "Original Data", xlab = "Value", ylab = "Frequency")
hist(simulated_data, main = "Simulated Data", xlab = "Value", ylab = "Frequency")

}
