% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish.R
\name{format_pairwise_plot}
\alias{format_pairwise_plot}
\title{Convert stats_PairwiseTest results as a dataframe for pairwise test plot}
\usage{
format_pairwise_plot(x, digit = 2)
}
\arguments{
\item{x}{stats_PairwiseTest object.}

\item{digit}{decimal places, default is 2.}
}
\value{
list of data.frames
}
\description{
Convert stats_PairwiseTest results as a dataframe for pairwise test plot
}
\examples{
library(tidyverse)
compare_res <- iris \%>\% stats_PairwiseTest(Species, .num_method="bonferroni_wilcox") \%>\% format_pairwise_plot(4)
iris \%>\% ggplot(aes(Species, Sepal.Length)) + geom_boxplot() +
    ggpubr::stat_pvalue_manual(compare_res$Sepal.Length \%>\% mutate(p.format = str_c(p.format, p.signif)), label = "p.format")
}
