% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish.R
\name{format_tab}
\alias{format_tab}
\title{Organize tibble for output using Officer package}
\usage{
format_tab(tab)
}
\arguments{
\item{tab}{tibble, output.}
}
\value{
Organized tibble viewed in Viewer window.
}
\description{
Organize tibble for output using Officer package
}
\examples{
read_docx() \%>\% body_add_flextable(format_tab(iris)) \%>\% print(target="./res.docx")
}
