\name{funTemp}
\alias{funTemp}
\title{analysis template}
\usage{
  funTemp(fun = "fun", folder = "/R")

  funTemp(fun = "fun", folder = "/R")
}
\arguments{
  \item{file}{a string of a file name}

  \item{fun}{a string of a function name}

  \item{folder}{a relative reference path to a destination
  folder (leave it empty for the current directory)}
}
\value{
  file.R

  fun.R
}
\description{
  create a analysis file template in the working directory

  create a function template in the working directory
}

