% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_shap.R
\name{get_model_predictions}
\alias{get_model_predictions}
\title{Get predictions from various model types}
\usage{
get_model_predictions(model, newx, model_type)
}
\description{
Get predictions from various model types
}
