% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_next_date.R
\name{get_next_date}
\alias{get_next_date}
\title{Get next date for a time series object}
\usage{
get_next_date(y)
}
\arguments{
\item{y}{A time series object}
}
\description{
Get next date for a time series object
}
\examples{

get_next_date(AirPassengers)
get_next_date(lynx)
get_next_date(USAccDeaths)

}
