% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_package_available.R
\name{is_package_available}
\alias{is_package_available}
\title{Check if a package is available}
\usage{
is_package_available(pkg_name)
}
\arguments{
\item{pkg_name}{A package name}
}
\value{
A logical value
}
\description{
Check if a package is available
}
\examples{

misc::is_package_available("dplyr")

}
