% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_whole_number.R
\name{is_wholenumber}
\alias{is_wholenumber}
\title{Check if a number is a whole number}
\usage{
is_wholenumber(x, tol = .Machine$double.eps^0.5)
}
\arguments{
\item{x}{A number}

\item{tol}{A tolerance level}
}
\value{
A logical value
}
\description{
Check if a number is a whole number
}
\examples{

is_wholenumber(1)
is_wholenumber(1.1)
is_wholenumber(1L)

}
