% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kl_div.R
\name{KL_divergence_hist}
\alias{KL_divergence_hist}
\title{Function to calculate KL divergence for continuous distributions using histograms}
\usage{
KL_divergence_hist(P, Q)
}
\arguments{
\item{P}{Numeric vector representing the empirical distribution}

\item{Q}{Numeric vector representing the theoretical distribution}
}
\value{
KL divergence between P and Q
}
\description{
Function to calculate KL divergence for continuous distributions using histograms
}
\examples{

P <- c(0.2, 0.3, 0.5)
Q <- c(0.1, 0.4, 0.5)
misc::KL_divergence_hist(P, Q)

}
