% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_shap.R
\name{local_linear_shap}
\alias{local_linear_shap}
\title{LIME-style local linear approximation using gradients}
\usage{
local_linear_shap(f, X_train, X_new, n_samples = 1000, sigma = 0.1, h = 1e-05)
}
\description{
LIME-style local linear approximation using gradients
}
