% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_scaled_pinball_loss.R
\name{mspl}
\alias{mspl}
\title{Mean Scaled Pinball Loss for Quantile Forecasts}
\usage{
mspl(
  y_true,
  y_pred,
  probs = c(0.005, 0.025, 0.165, 0.25, 0.5, 0.75, 0.835, 0.975, 0.995)
)
}
\arguments{
\item{y_true}{Numeric vector of observed values}

\item{y_pred}{Numeric matrix or data.frame of predicted quantiles
(rows = observations, columns = quantiles)}

\item{probs}{Numeric vector of quantile levels (must be between 0 and 1)
Default \code{c(0.005, 0.025, 0.165, 0.25, 0.5, 0.75, 0.835, 0.975, 0.995)}}
}
\value{
Numeric scalar representing the MSPL
}
\description{
Calculates the Mean Scaled Pinball Loss (MSPL) for quantile forecasts,
scaled by the Mean Absolute Error (MAE) of the median forecast.
}
\examples{
y_true <- c(1, 2, 3, 4, 5)
y_pred <- matrix(c(0.5, 1.5, 2.5, 3.5, 4.5,
                   1.0, 2.0, 3.0, 4.0, 5.0,
                   1.5, 2.5, 3.5, 4.5, 5.5), ncol = 3)
mspl(y_true, y_pred, probs=c(0.25, 0.5, 0.75))
}
