% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_count.R
\name{na_count}
\alias{na_count}
\title{Count NA frequency in data}
\usage{
na_count(data)
}
\arguments{
\item{data}{a data frame}
}
\value{
a tibble
}
\description{
Count NA frequency in data
}
\examples{
\dontrun{
na_data <- data.frame(c1 = c(1, NA), c2 = c(NA, NA))
na_data \%>\% na_count()
}
}
