% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_viz.R
\name{na_viz}
\alias{na_viz}
\title{Visualize NA frequency in data}
\usage{
na_viz(data)
}
\arguments{
\item{data}{a data frame}
}
\value{
a ggplot object
}
\description{
\code{na_viz()} create a ggplot plot showing the percentage of NA in each column
}
\section{Acknowledgment}{

\code{na_viz()} is another name for the excellent \code{vis_miss()} of
\href{https://github.com/njtierney/naniar}{\code{{naniar}}}
}

\examples{
\donttest{
if (interactive()) {
  na_data <- data.frame(c1 = c(1, NA), c2 = c(NA, NA))
  na_data \%>\% na_viz()
}
}
}
\seealso{
Other missing-data:
\code{\link[=filter_na]{filter_na()}},
\code{\link[=na_count]{na_count()}},
\code{\link[=remove_columns_based_on_NA]{remove_columns_based_on_NA()}}
}
\concept{missing-data}
