% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np_stat_test.R
\name{np_stat_test}
\alias{np_stat_test}
\title{Non-parametric test for any statistic}
\usage{
np_stat_test(a_data, b_data, stat_fun = mean, n_sim = 1000,
  alternative = "two.sided", alpha = 0.05, plot = TRUE, ...)
}
\arguments{
\item{a_data}{Numeric vector of values from group a.}

\item{b_data}{Numeric vector of values from group b.}

\item{stat_fun}{Function that calculates desired statistic.}

\item{n_sim}{Number of times to sample groups and calculate difference.}

\item{alternative}{Setting up alternative hypothesis:
"two.sided" meaning stat(A) != stat(B),
"less" meaning stat(A) < stat(B),
"greater" meaning stat(A) > stat(B)}

\item{alpha}{statistical significance}

\item{plot}{(logical) Plot the distribution of differences in mean if TRUE.}

\item{...}{Any additional parameters that stat_fun needs.}
}
\value{
returns a named list containing:
\item{diff}{all the statistics differences}
\item{observed}{observed statistics difference}
\item{n_sim}{number of simulations ran}
\item{p_value}{two sided p-value for the test}
}
\description{
Performs a non parametric test for any statistics by bootstraping. Combines all values and samples
vectors of length a_data, b_data. Then calculates differences in statistics to get distribution
of differences in statatistics. Then compares observed difference with that distribution.
}
\details{
H0: No difference in statistics
}
\examples{
x = rnorm(300, mean = 5, sd = 1)
y = rnorm(200, mean = 4, sd = 1.5)
y2 = c(y, 110)

np_stat_test(x, y, mean)
np_stat_test(x, y2, mean)
np_stat_test(x, y, mean, trim = 0.1)
np_stat_test(x, y, median)

}
\author{
Elio Bartoš
}
