% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_hot.R
\name{one_hot_encode}
\alias{one_hot_encode}
\title{One-hot encoding}
\usage{
one_hot_encode(y)
}
\arguments{
\item{y}{A vector of class labels}

\item{n_classes}{The number of classes}
}
\value{
A matrix of one-hot encoded labels
}
\description{
One-hot encoding
}
\examples{

y <- as.factor(c(1, 2, 1, 1, 2))
misc::one_hot_encode(y)

}
