% Generated by roxygen2 (4.0.2): do not edit by hand
\name{or}
\alias{or}
\title{Odds Ratio}
\usage{
or(x, ci = 0.95)
}
\arguments{
\item{x}{numeric matrix, contigency table}

\item{ci}{numeric value,confidence interval}
}
\value{
a list, with tables and odds ratios
}
\description{
calculate OR with CI
}
\examples{
x <- matrix(c(7,23,41,43,27,9), nrow=3)
dimnames(x) <- list(c('trt1','trt2','trt3'), c('pass','fail'))
or(x)
}

