% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametrized_query.R
\name{parametrized_query}
\alias{parametrized_query}
\title{Parametrize a SQL query}
\usage{
parametrized_query(query, params)
}
\arguments{
\item{query}{string representing your query with x1x, x2x, ... on places where you want to put your parameters}

\item{params}{list or vector of strings to be placed in a query on places x1x, x2x,.. respectively.}
}
\value{
A parametrised query ready to be sent to the database
}
\description{
In a query string put x1x, x2x and so on characters where you want to insert your strings. Provide a list of
of strings and those values will be replaced and query will be ready to be sent to database.
}
\examples{
query = "
select date, sum(revenue) as revenue
  from my_sample_table
  where date between 'x1x' and 'x2x'
  group by 1
  order by 1;
"
p_query = parametrized_query(query, c('2018-09-01', as.character(Sys.Date())))
cat(p_query)

}
\author{
Elio Bartoš
}
