% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformalize.R
\name{predict.conformalize}
\alias{predict.conformalize}
\title{Predict method for conformalize objects}
\usage{
\method{predict}{conformalize}(
  object,
  newdata,
  predict_func = predict,
  level = 0.95,
  method = c("splitconformal", "kde", "surrogate", "bootstrap"),
  n_sim = 250L,
  seed = 123L,
  ...
)
}
\arguments{
\item{object}{A conformalize object.}

\item{newdata}{A data frame or matrix of new data for prediction.}

\item{level}{The confidence level for the prediction intervals. Default is 0.95.}

\item{method}{The method to use for prediction intervals. Options are "split" or "simulation".}

\item{n_sim}{The number of simulations to perform if using the "simulation" method. Default is 1000.}

\item{seed}{An optional seed for reproducibility.}

\item{...}{Additional arguments to pass to the predict function.}
}
\value{
A matrix with predictions and prediction intervals.
}
\description{
This function generates prediction intervals for new data using the fitted conformal model.
}
