% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefer.R
\name{prefer}
\alias{prefer}
\title{Defines preferred functions from conflicts between namespaces}
\usage{
prefer(pkg_name, quiet = TRUE)
}
\arguments{
\item{pkg_name}{a atomic vector with package names}

\item{quiet}{If warnings should be displayed. Default is TRUE}
}
\description{
Defines preferred functions from conflicts between namespaces
}
\examples{
\dontrun{
# prefer `{dplyr}` functions over `{stats}`
prefer("dplyr")
}
}
