% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish.R
\name{publish}
\alias{publish}
\alias{publish.desc_baseline}
\alias{publish.stats_SingleTest}
\alias{publish.stats_PairwiseTest}
\alias{publish.stats_Correlation}
\title{Convert some objects from 'misc' package to a dataframe what can be considered as publish tables}
\usage{
publish(x, ...)

\method{publish}{desc_baseline}(x, digit = 2)

\method{publish}{stats_SingleTest}(x, digit = 2)

\method{publish}{stats_PairwiseTest}(x, digit = 2)

\method{publish}{stats_Correlation}(x, digit = 2)
}
\arguments{
\item{x}{desc_baseline, stats_SingleTest or stats_PairwiseTest object.}

\item{digit}{decimal places, default is 2.}
}
\value{
data.frame
}
\description{
Convert some objects from 'misc' package to a dataframe what can be considered as publish tables
}
\examples{

}
