% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sheet_then_save_csv.R
\name{read_all_sheets_then_save_csv}
\alias{read_all_sheets_then_save_csv}
\title{Read and save all excel sheets and save them to a CSV file}
\usage{
read_all_sheets_then_save_csv(path_to_xlsx, dir_to_save = NULL)
}
\arguments{
\item{path_to_xlsx}{a character vector with path to the excel file}

\item{dir_to_save}{a character vector with the path to save the csv files.
Default is NULL and save the csv files in the "data/temp/extracted_sheets"
if it exists.}
}
\value{
A list (one element per sheet), each the return value of
\code{\link[=read_sheet_then_save_csv]{read_sheet_then_save_csv()}} for that sheet (invisibly \code{NULL} per call).
}
\description{
\code{read_all_sheets_then_save_csv()} just loops \code{read_sheet_then_save_csv()} over
the available excel sheets and save them in \code{data/temp/extracted_sheets}
}
\section{Acknowledgment}{

See: \code{\link{read_sheet_then_save_csv}}
}

\examples{
\donttest{
if (interactive()) {
  # read and into a csv
  misc::create_dirs("ma-box")
  xlsx_file <- system.file("xlsx-examples", "mtcars_workbook_001.xlsx", package = "misc")
  read_all_sheets_then_save_csv(
    path_to_xlsx = xlsx_file,
    dir_to_save = "ma-box"
  )
}
}
}
\seealso{
Other excel-import:
\code{\link[=read_all_xlsx_then_save_csv]{read_all_xlsx_then_save_csv()}},
\code{\link[=read_sheet_then_save_csv]{read_sheet_then_save_csv()}}
}
\concept{excel-import}
