\name{readLines2}
\alias{readLines2}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Read Text Lines from a Connection~
}
\description{
Import text files
}
\usage{
readLines2(fname)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{fname}{
Name (and path) of text file
}
}
\details{
See \href{http://mlt-thinks.blogspot.nl/2011/08/faster-files-in-r.html}{this blog post} for the original function and explanation
}
\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
Paulo van Breugel
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{

}
