% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_geom_holes.R
\name{remove_geom_holes}
\alias{remove_geom_holes}
\title{Remove holes from polygon geometries}
\usage{
remove_geom_holes(x, max_area = 0)
}
\arguments{
\item{x}{An \code{sf}, \code{sfc}, or \code{sfg} object of type \code{POLYGON} or
\code{MULTIPOLYGON}.}

\item{max_area}{Maximum area of holes to remove, in the units of \code{x} (or
square metres for geographic coordinates). Default \code{0} removes all holes.}
}
\value{
An object of the same class as \code{x}, with selected holes removed.
}
\description{
Removes polygon holes from an \code{sf}, \code{sfc}, or \code{sfg} object. With a positive
\code{max_area}, only holes smaller than the threshold are removed.
}
\note{
For a faster alternative when \code{max_area} is unnecessary, see
\code{sfheaders::sf_remove_holes()}:
\url{https://github.com/dcooley/sfheaders}.
}
\section{Acknowledgment}{

This implementation is adapted from
\href{https://github.com/michaeldorman/nngeo}{\code{nngeo::st_remove_holes()}},
written by Michael Dorman and distributed under the MIT License.
}

\references{
Adapted from the StackOverflow answer by user \code{lbusett}:
\url{https://stackoverflow.com/questions/52654701/removing-holes-from-polygons-in-r-sf}.
}
\seealso{
Other geo-tools:
\code{\link[=fix_invalid_geometries]{fix_invalid_geometries()}},
\code{\link[=intersect_filter_touch]{intersect_filter_touch()}},
\code{\link[=intersect_mask_filter_area]{intersect_mask_filter_area()}},
\code{\link[=quick_map]{quick_map()}},
\code{\link[=view_mapview_from_path]{view_mapview_from_path()}}
}
\concept{geo-tools}
