% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformalize.R
\name{residuals.conformalize}
\alias{residuals.conformalize}
\title{Residuals method for conformalize objects}
\usage{
\method{residuals}{conformalize}(object, ...)
}
\arguments{
\item{object}{A conformalize object.}

\item{...}{Additional arguments (not used).}
}
\value{
A numeric vector of residuals.
}
\description{
This function extracts the residuals from a conformalize object.
}
