% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_zero_cols.R
\name{rm_zero_cols}
\alias{rm_zero_cols}
\title{Removing columns containing only zeros}
\usage{
rm_zero_cols(X)
}
\arguments{
\item{X}{A matrix or data frame}
}
\value{
A matrix or data frame
}
\description{
Removing columns containing only zeros
}
\examples{

X <- matrix(c(1, 0, 3, 0, 5, 0, 0, 0), nrow = 2)
print(misc::rm_zero_cols(X))

}
