% Generated by roxygen2 (4.0.2): do not edit by hand
\name{rp}
\alias{rp}
\title{Rounded print}
\usage{
rp(x, n = 4)
}
\arguments{
\item{x}{numeric vector}

\item{n}{integer value, this should
be larger than zero}
}
\value{
printed output
}
\description{
floating point printing
}
\examples{
rp(1/seq(10))
rp(1/seq(10), 6)
}

