% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_plot.R
\name{save_plot}
\alias{save_plot}
\title{Save a ggplot figure}
\usage{
save_plot(
  object,
  filename = NULL,
  dir_to_save = NULL,
  width = NA,
  height = NA,
  format = NULL,
  units = NULL,
  dpi = NULL,
  overwrite = FALSE,
  trim = FALSE
)
}
\arguments{
\item{object}{a ggplot object}

\item{filename}{a character vector with the name of the file to save. Default is NULL and saves with the name of the object}

\item{dir_to_save}{a character vector with the name of the directory to save}

\item{width}{a numerical vector with the width of the figure}

\item{height}{a numerical vector with the height of the figure}

\item{format}{a character vector with format of the figure. Can "jpeg", "tiff", "png" (default), or "pdf"}

\item{units}{a character vector with the units of the figure size. Can be "in", "cm" (default), or "mm"}

\item{dpi}{a numerical vector with the resolution of the figure. Default is 300}

\item{overwrite}{logical}

\item{trim}{logical}
}
\value{
No return value, called for side effects (writes a graphics file
via \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}, and optionally calls \code{\link[=trim_fig]{trim_fig()}}).
}
\description{
\code{save_plot()} wraps \code{ggplot2::ggsave()} and offer option to remove white
spaces around figures (creates a additional file in \code{output/figures/trim};
uses \code{\link{trim_fig}})
}
\section{Acknowledgment}{

\code{save_plot()} is derived from
\href{https://github.com/globeandmail/startr/blob/fff446f5a07a67f565a7bae887f0cdd24c808cdb/R/utils.R#L157}{\code{write_plot()}},
available in the excellent
\href{https://github.com/globeandmail/startr}{\code{start}} project template
}

\examples{
\donttest{
if (interactive()) {
  library(misc)
  ipak(c("ggplot2", "dplyr"))
  create_dirs()
  p <- mtcars \%>\%
    ggplot() +
    aes(x = mpg, y = cyl) +
    geom_point()
  save_plot(p)
}
}
}
\seealso{
Other save-output:
\code{\link[=save_temp_data]{save_temp_data()}},
\code{\link[=trim_fig]{trim_fig()}}
}
\concept{save-output}
