% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_matrix.R
\name{scale_matrix}
\alias{scale_matrix}
\title{Scale matrix}
\usage{
scale_matrix(X, X_mean = NULL, X_sd = NULL)
}
\arguments{
\item{X}{A matrix}

\item{X_mean}{Mean of each column}

\item{X_sd}{Standard deviation of each column}
}
\value{
A list containing the scaled matrix, mean of each column, and standard deviation of each column
}
\description{
Scale matrix
}
\examples{

X <- matrix(c(1, 2, 3, 4, 5, 6), nrow = 2)
(X_scaled <- misc::scale_matrix(X))
(X_scaled <- misc::scale_matrix(X, X_mean = colMeans(X), X_sd = apply(X, 2, stats::sd)))
print(colMeans(X_scaled$X))
print(apply(X_scaled$X, 2, stats::sd))

}
