% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_lintr.R
\name{setup_lintr}
\alias{setup_lintr}
\title{Setup a config file for static code analysis}
\usage{
setup_lintr(exclude_file = NULL, exclude_path = NULL)
}
\arguments{
\item{exclude_file}{character a character vector containing the files that
should be excluded from static code analysis.}

\item{exclude_path}{character a character vector containing the paths that
should be excluded from static code analysis.}
}
\value{
none
}
\description{
This function create a config file (\code{.lintr}) in root of the user's project
based on static code analysis performed by the \code{{lintr}}.
}
\details{
\code{setup_lint} creates a file (\code{.lintr}) in the root of the user project,
which lists and counts the warnings, syntax errors, and possible semantic
issues provided by the static code analysis of the \code{{lintr}}. In the current
form (i.e., right after its creation), the \code{.lintr} file will block all the
problems and hints founded by the static code analysis. Users need to edit
the file manually, changing the variable values from NULL to the value used
by each type of linter listed. For example, the default value of the
\code{line_length_linter} linter is 80, but users can change it to 120. Please,
check the \href{https://github.com/jimhester/lintr/blob/master/README.md}{README}
of the \code{{lintr}} for further information.
}
\examples{
\dontrun{
setup_lintr(exclude_file = "manuscript.Rmd", exclude_path = c("data", "sources"))
}
}
