% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformalize.R
\name{simulate.conformalize}
\alias{simulate.conformalize}
\title{simulate method for conformalize objects}
\usage{
\method{simulate}{conformalize}(
  object,
  newdata,
  method = c("kde", "surrogate", "bootstrap"),
  n_sim = 250L,
  seed = NULL,
  ...
)
}
\arguments{
\item{object}{A object of class \code{conformalize}.}

\item{newdata}{A data frame or matrix of new data for prediction.}

\item{method}{The method to use for prediction intervals. Options are "split" or "simulation".}

\item{n_sim}{The number of simulations to perform if using the "simulation" method. Default is 1000.}

\item{seed}{An optional seed for reproducibility.}

\item{...}{Additional arguments to pass to the predict function.}
}
\value{
A matrix with predictions and prediction intervals.
}
\description{
This function generates simulations for new data using the fitted conformal model.
}
