% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_dataframe.R
\name{sort_df}
\alias{sort_df}
\title{Sort data frame}
\usage{
sort_df(df, by, decreasing = FALSE)
}
\arguments{
\item{df}{data frame}

\item{by}{column to sort by}

\item{decreasing}{logical. Should sorting be decreasing?}
}
\value{
A sorted data frame
}
\description{
Sort data frame
}
\examples{

df <- data.frame(a = c(2, 4, 3), b = c(3, 5, 1))
misc::sort_df(df, "a")
misc::sort_df(df, "b", decreasing = TRUE)

}
