% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_data.R
\name{split_data}
\alias{split_data}
\title{Split a dataset}
\usage{
split_data(y, p = 0.5, seed = 123, type_split = c("stratify", "sequential"))
}
\arguments{
\item{y}{A vector of labels}

\item{p}{A proportion of the dataset to split}

\item{seed}{An integer to set the seed}

\item{type_split}{A character string specifying the type of split}
}
\value{
A vector of indices
}
\description{
Split a dataset
}
\examples{

set.seed(123)
(y <- rnorm(10))
misc::split_data(y, 0.5)
misc::split_data(y, 0.5, type_split = "sequential")

}
