% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_Correlation.R
\name{stats_Correlation}
\alias{stats_Correlation}
\title{Correlation Analysis}
\usage{
stats_Correlation(dat, ..., .method = "spearman")
}
\arguments{
\item{dat}{DataFrame.}

\item{...}{if NULL, use all variables, otherwise the variables here will be calculated.}

\item{.method}{Default "spearman", "pearson".}
}
\value{
An stats_Correlation object.
}
\description{
Correlation Analysis
}
