% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_MultiTest.R
\name{stats_MultiTest}
\alias{stats_MultiTest}
\title{Multiple elements Analysis for data.frame}
\usage{
stats_MultiTest(formula, dat, .digit = 2)
}
\arguments{
\item{formula}{formula, like \code{glm}}

\item{dat}{data.frame}

\item{.digit}{Integer, significant number of results retained.}
}
\value{
tibble, the element is character
}
\description{
Multiple elements Analysis for data.frame
}
\examples{
stats_MultiTest(mpg~., mtcars)
}
