% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_PairwiseTest.R
\name{stats_PairwiseTest}
\alias{stats_PairwiseTest}
\alias{stats_PairwiseTest.data.frame}
\alias{stats_PairwiseTest.stats_SingleTest}
\title{Pairwise Testing}
\usage{
stats_PairwiseTest(x, ...)

\method{stats_PairwiseTest}{data.frame}(
  dat,
  group,
  ...,
  .num_method = "auto",
  .fct_method = "auto",
  .adjust_method = "holm",
  .check_NE_args = list()
)

\method{stats_PairwiseTest}{stats_SingleTest}(
  x,
  .adjust_method = "holm",
  .just_for_sign = FALSE,
  .sign_thre = 0.05
)
}
\arguments{
\item{x}{\code{stats_SingleTest} object.}

\item{...}{One or more unquoted expressions separated by commas}

\item{dat}{data.frame.}

\item{group}{The group variable unquoted name}

\item{.num_method}{Character. "auto", "tukey", "t", "wilcox".}

\item{.fct_method}{Character. "auto", "chisq", "fisher".}

\item{.adjust_method}{Character. "holm", "bonferroni", "hochberg", "hommel".}

\item{.check_NE_args}{Named list. If \code{.num_method} is "auto", these are the parameters of \code{.check_NE}.
The default parameters are \code{list(.N = TRUE, .E = TRUE, .Nthre = 0.1, .Ethre = 0.1, .Ntest_object="every")}.}

\item{.just_for_sign}{Logical value. If \code{TRUE}, just test for variables that be tested as significant in \code{stats_SingleTest}.}

\item{.sign_thre}{Float threshold used to decide which variable is significant. Default 0.05.}
}
\value{
\code{stats_PairwiseTest} object.
\describe{
  \item{main}{This is a list whose items are 2-length lists which contains the pairwise test results.
    The two items are \code{p_value}(named matrix) and \code{method}.}
  \item{SingleTest}{If the input is \code{stats_SingleTest} object, this is it.}
  \item{group, vars}{Stored original data.}
}
}
\description{
This function is used to compare all pairs of categories of \code{group} variable.
This pairwise test methods used contain bonferroni adjust of t, wilcox and chisq, and tukey method
for continues variables.
}
\section{Methods (by class)}{
\itemize{
\item \code{data.frame}: Pairwise Test for data.frame.

\item \code{stats_SingleTest}: Pairwise Test for \code{stats_SingleTest} object. The test method will be choose
by the \code{x$test_methods}. \code{"anova"} uses \code{"tukey"}, \code{"kwh"} uses \code{"wilcox"},
\code{"chisq"} uses \code{"chisq"}, and \code{"fisher"} uses \code{"fisher"}. The \code{"wilcox"}, \code{"chisq"} and \code{"fisher"}
will be adjusted by \code{".adjust_method"}.
}}

\examples{
stats_PairwiseTest(iris, Species)
}
