% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_SingleTest.R
\name{stats_SingleTest}
\alias{stats_SingleTest}
\title{Single Null Hypothesis Significant Test (NHST) for selected variables of data.frame}
\usage{
stats_SingleTest(
  dat,
  group,
  ...,
  .desc = TRUE,
  .num_method = "auto",
  .fct_method = "auto",
  .check_NE_args = list(),
  .desc_baseline_args = list()
)
}
\arguments{
\item{dat}{data.frame}

\item{group}{The group variable unquoted name.}

\item{...}{One or more unquoted expressions separated by commas. If none, use every variables else.}

\item{.desc}{Logical, whether calculate \code{desc_baseline} object.}

\item{.num_method}{Character, "auto", "t", "anova", "wilcox", "kruskal". If \code{nlevels(group)} > 2 and "t" or
"wilcox" are chosen, a warning will be throwed and the parameter will be replaced by "anova" or "kruskal". Default
is "auto", the test method of numeric variables will be decided by \code{.check_NE}. The "t" means t test with Welch
(or Satterthwaite) approximation of the degrees of freedom.}

\item{.fct_method}{Character, "auto", "chisq", "fisher". Default is "auto", decided by \code{.check_chisq}.}

\item{.check_NE_args}{Named list. If \code{.num_method} is "auto", these are the parameters of \code{.check_NE}.
The default parameters are \code{list(.N = TRUE, .E = TRUE, .Nthre = 0.1, .Ethre = 0.1, .Ntest_object="every")}.}

\item{.desc_baseline_args}{Named list. If \code{.desc=TRUE}, these are the parameters of \code{.desc_baseline}.
The default parameters are \code{.num_desc_type = "auto", .num_margin = TRUE, .fct_margin = "all", .fct_prob = "col", .check_NE_args = list()}.
Because the parameters will pass into \code{desc_baseline}, \code{.check_NE_args} will use the default values of
\code{desc_baseline}.}
}
\value{
stats_SingleTest object.
\describe{
    \item{main}{This is a list whose items are 3-length lists which contains the NHST results. The three items
    are \code{statistic}, \code{p_value} and \code{method}.}
    \item{test_methods}{Named vector. Its items are the \code{method} of every variable used for NHST.}
    \item{desc}{The \code{desc_baseline} object. See \code{\link{desc_baseline}}.}
    \item{vars, group}{Stored original data.}
}
}
\description{
Single Null Hypothesis Significant Test (NHST) for selected variables of data.frame
}
\examples{
stats_SingleTest(iris, Species)
stats_SingleTest(iris, Species, .desc = FALSE)
}
