% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeit.R
\name{timeit}
\alias{timeit}
\title{Timing an expression}
\usage{
timeit(expr, times = 1, ...)
}
\arguments{
\item{expr}{an R expression}

\item{times}{number of repetitions}

\item{...}{additional arguments passed to \code{\link{base::eval}}}
}
\value{
the elapsed time in seconds
}
\description{
Timing an expression
}
\examples{

timeit(1 + 1)
timeit(1 + 1, times = 10)

}
