% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_fig.R
\name{trim_fig}
\alias{trim_fig}
\title{Remove white spaces around figures}
\usage{
trim_fig(figure_path, overwrite = FALSE)
}
\arguments{
\item{figure_path}{a character vector with path of the figure}

\item{overwrite}{logical}
}
\value{
No return value, called for side effects (writes a trimmed image
file under a \verb{trim/} subdirectory via \code{\link[magick:image_write]{magick::image_write()}}).
}
\description{
\code{trim_fig()} just remove white spaces around a figure and save it into the
\code{trim} folder (maintain the original figure untouchable)
}
\section{Acknowledgment}{

\code{trim_fig()} wraps the excellent \code{image_trim()} of
\href{https://github.com/ropensci/magick}{\code{{magick}}}
}

\examples{
\donttest{
if (interactive()) {
  library(misc)
  ipak(c("ggplot2", "dplyr"))
  create_dirs()
  p <- mtcars \%>\%
    ggplot() +
    aes(x = mpg, y = cyl) +
    geom_point()
  save_plot(p)
  trim_fig("output/figures/p.png")
}
}
}
\seealso{
Other save-output:
\code{\link[=save_plot]{save_plot()}},
\code{\link[=save_temp_data]{save_temp_data()}}
}
\concept{save-output}
