% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_shap.R
\name{validate_shap_decomposition}
\alias{validate_shap_decomposition}
\title{Validate SHAP decomposition}
\usage{
validate_shap_decomposition(shap_values, base_value, predictions)
}
\description{
Validate SHAP decomposition
}
