% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vlookup.R
\name{vlookup}
\alias{vlookup}
\title{VLOOKUP}
\usage{
vlookup(this, df, key, value)
}
\arguments{
\item{this}{The value to look up}

\item{df}{A data frame}

\item{key}{The column to look up}

\item{value}{The column to return}
}
\value{
The value in the `value` column corresponding to the `key` column
}
\description{
A simple implementation similar to the VLOOKUP function in Excel.
}
\examples{

df <- data.frame(key = c("a", "b", "c"), value = c(1, 2, 3))
print(misc::vlookup("b", df, "key", "value"))
 
}
