% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_excel.R
\name{write_excel}
\alias{write_excel}
\title{Copy data frame to clipboard}
\usage{
write_excel(x, row.names = FALSE, col.names = TRUE, ...)
}
\arguments{
\item{x}{Data frame to be copied.}

\item{row.names}{(boolean) Should row names be copied? Default: FALSE}

\item{col.names}{(boolean) should column names be copied? Default: TRUE}

\item{...}{parameters forwarded to write.table}
}
\value{
None
}
\description{
Copies data frame to clipboard. Usually used when you want to copy data frame to Excel. Just run
function with desired df and press Paste or CTRL + V to paste it in the desired place. This function is like
pressing CTRL + C on a dataframe. It should work on MAC and Windows.
}
\examples{
x = data.frame(var1 = c(0, 1), var2 = c(3, 4))
write_excel(x)

}
\author{
Elio Bartoš
}
