% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_gitignore.R
\name{add_gitignore}
\alias{add_gitignore}
\title{Add a gitignore to the project root}
\usage{
add_gitignore(type = "r")
}
\arguments{
\item{type}{a character vector with the language to be ignored}
}
\description{
Add a gitignore to the project root
}
\examples{
\dontrun{
# if your project do not have a gitignore, take it:
add_gitignore()
}
}
