% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_words_ptbr.R
\name{combine_words_ptbr}
\alias{combine_words_ptbr}
\title{Combine words using ptbr rules}
\usage{
combine_words_ptbr(words, sep = NULL, last = NULL)
}
\arguments{
\item{words}{a character vector with words to combine}

\item{sep}{a character with the separator of the words. Default is NULL and insert ", "}

\item{last}{a character vector with the last separator of the words. Default is NULL and insert " e "}
}
\value{
a character vector
}
\description{
\code{combine_words_ptbr()} collapse words using ptbr rules. This function
differ from \href{https://github.com/yihui/knitr/blob/2e08d5a9cb0a3b83cd73fd30d507a18676b0c4a4/R/utils.R#L830}{\code{knitr::combine_words()}}
which uses oxford commas.
}
\section{Acknowledgment}{

\code{combine_words_ptbr()} uses \href{https://cran.r-project.org/web/packages/glue/vignettes/transformers.html}{transformers}
available in the excellent \href{https://github.com/tidyverse/glue}{\code{{glue}}} package
}

\examples{
\dontrun{
misc::ipak(c("dplyr", "broom", "glue"))

# showing significant variables
mtcars \%>\%
  select(mpg, cyl, carb, wt) \%>\%
  lm(mpg ~ cyl + carb + wt, data = .) \%>\%
  tidy() \%>\%
  filter(p.value <= 0.05, term != "(Intercept)") \%>\%
  pull(term) \%>\% {
    paste("Variáveis significativas:", combine_words_ptbr(.))
  }

# using in an ordinary text
feira <- c("banana", "maça", "pepino", "ovos")
glue("Por favor, compre: {combine_words_ptbr(feira)}")
}
}
