% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dirs.R
\name{create_dirs}
\alias{create_dirs}
\title{Create data, output and R directories}
\usage{
create_dirs(dirs = NULL)
}
\arguments{
\item{dirs}{a character vector with the directory names. Default is NULL and create \code{data/raw}, \code{data/clean}, \code{data/temp}, \code{output/figures}, \code{output/results}, \code{output/supp}, and \code{R}}
}
\description{
Create data, output and R directories
}
\examples{
\dontrun{
# create a single directory
create_dirs("myfolder")
# create the default directories
create_dirs()
# see the resulting tree
fs::dir_tree()
}
}
