% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_na.R
\name{filter_na}
\alias{filter_na}
\title{Easily filter NA values from data frames}
\usage{
filter_na(data, type = c("any", "all"))
}
\arguments{
\item{data}{a data frame or tibble}

\item{type}{a character vector indicating which type of NA-filtering must be done. If type = "any",
\code{filter_na()} will filter any NA values present in the data frame. If type = "all", \code{filter_na} will
filter only rows which all columns has NA values.}
}
\value{
a tibble object
}
\description{
\code{filter_na()} just wrap \code{{dplyr}} functions in a more
convenient way, IMO.
}
\examples{
\dontrun{
nice_data <- data.frame(c1 = c(1, NA), c2 = c(NA, NA))
nice_data \%>\%
  filter_na("all")
nice_data \%>\%
  filter_na("any")
}
}
