% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_value.R
\name{get_value}
\alias{get_value}
\title{Get a value from function defined by points}
\usage{
get_value(df, x_value, method = "constant")
}
\arguments{
\item{df}{Data frame with two columns, first column representing x values and second y values.}

\item{x_value}{Value for which we want to calculate this function.}

\item{method}{('constant', 'linear') Should function be step function (constant between points), or interpolated linearly.}
}
\value{
Value at x_value point.
}
\description{
For a function defined by points and then interpolated as constant or linear function get a value
for specific x. Function points should be provided in data frame with 2 columns: x and y values.
}
\examples{
df = tribble(
~x, ~y,
 1, 1,
 5, 5,
 10, 10
)

get_value(df, 2)           # returns 1
get_value(df, 2, "linear") # returns 2
get_value(df, 7)           # returns 5
get_value(df, 7, "linear") # returns 7

}
\author{
Elio Bartoš
}
