% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_viz.R
\name{na_viz}
\alias{na_viz}
\title{Vizualize NA frequency in data}
\usage{
na_viz(data)
}
\arguments{
\item{data}{a data frame}
}
\value{
a ggplot object
}
\description{
Vizualize NA frequency in data
}
\examples{
\dontrun{
na_data <- data.frame(c1 = c(1, NA), c2 = c(NA, NA))
na_data \%>\% na_viz()
}
}
