% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np_mean_test.R
\name{np_mean_test}
\alias{np_mean_test}
\title{Non-parametric mean test}
\usage{
np_mean_test(a_data, b_data, n_sim = 1000, plot = TRUE)
}
\arguments{
\item{a_data}{a numeric vector of values from group a}

\item{b_data}{a numeric vector of values from group b}

\item{n_sim}{number of times to sample groups and calcutale difference}

\item{plot}{(logical) plot the distribution of differences in mean if TRUE}
}
\value{
returns a list containing
\item{name a}{description a}
\item{name b}{description b}
}
\description{
Performs a non parametric mean test by bootstraping. Combines all values and samples
vectors of length a_data, b_data. Then calculates differences in mean to get distribution
of differences in mean. Then compares observed difference with that distribution.
}
\author{
Elio Bartos
}
