% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parfor.R
\name{parfor}
\alias{parfor}
\title{Sequential or parallel for loop.}
\usage{
parfor(
  what,
  args,
  cl = NULL,
  combine = c,
  errorhandling = c("stop", "remove", "pass"),
  verbose = FALSE,
  show_progress = TRUE,
  export = NULL,
  ...
)
}
\arguments{
\item{what}{A function.}

\item{args}{A list of arguments.}

\item{cl}{Number of cores to use. If \code{NULL}, the loop will be sequential. It -1, the number of cores will be detected automatically.}

\item{combine}{A function to combine the results.}

\item{errorhandling}{A character string specifying how to handle errors. Possible values are \code{"stop"}, \code{"remove"}, and \code{"pass"}.}

\item{verbose}{A logical indicating whether to print progress.}

\item{show_progress}{A logical indicating whether to show a progress bar.}

\item{export}{A list of objects to export to the workers.}

\item{...}{Additional arguments to pass to \code{what} for \code{\link{foreach::foreach}} (excludind \code{.combine}, \code{.errorhandling}, \code{.options.snow}, \code{.verbose}, and \code{.export}).}
}
\value{
A list of results.
}
\description{
Sequential or parallel for loop.
}
\examples{

# Sequential
print(misc::parfor(function(x) x^2, 1:10))

# Parallel
print(misc::parfor(function(x) x^2, 1:10, cl = 2))

}
