% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_map.R
\name{quick_map}
\alias{quick_map}
\title{Create maps quickly}
\usage{
quick_map(region = NULL, type = NULL)
}
\arguments{
\item{region}{character string or atomic vector containing countries names ou continents. Default is \code{NULL}.}

\item{type}{character string informing map type. Can be \code{"sf"} or \code{"ggplot"}}
}
\value{
a ggplot object
}
\description{
Create maps quickly
}
\examples{
\dontrun{
# plot a world map
quick_map()
# plot a new world map
quick_map(region = "Americas", type = "sf")
# using ggplot
quick_map(region = "Americas", type = "ggplot")
}
}
