% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_map.R
\name{quick_map}
\alias{quick_map}
\title{Create maps quickly}
\usage{
quick_map(region = NULL, type = NULL)
}
\arguments{
\item{region}{character string or atomic vector containing countries names ou continents. Default is \code{NULL}.}

\item{type}{character string informing map type. Can be \code{"sf"} or \code{"ggplot"}}
}
\value{
a ggplot object
}
\description{
\code{quick_map()} allows the creation of maps quickly using \code{{ggplot2}}. For this
reason, the resulting map is fully editable through \code{{ggplot2}} layers.
}
\section{Acknowledgment}{

\code{quick_map()} depends heavily on the data available by
the \href{https://github.com/ropensci/rnaturalearth}{\code{{rnaturalearth}}}
package. In this sense, \code{quick_map()} uses a wide and dirty filtering of
this data to create the map.
}

\examples{
\dontrun{
# plot a world map
quick_map()
# plot a new world map
quick_map(region = "Americas", type = "sf")
# using ggplot
quick_map(region = "Americas", type = "ggplot")
# edit using ggplot2 layers
quick_map() +
  theme_void() +
  geom_sf(fill = "white")
}
}
