% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sheet_then_save_csv.R
\name{read_all_xlsx_then_save_csv}
\alias{read_all_xlsx_then_save_csv}
\title{Read all sheets from all excel files and save into CSV files}
\usage{
read_all_xlsx_then_save_csv(path_to_xlsx)
}
\arguments{
\item{path_to_xlsx}{a character vector with the path to excel file}
}
\description{
Following the same principle of \code{\link{read_all_sheets_then_save_csv}}
\code{read_all_xlsx_then_save_csv()} just loop \code{read_all_sheets_then_save_csv()} over
all available xls\link{x} files
}
\section{Acknowledgment}{

See: \code{\link{read_sheet_then_save_csv}}
}

\examples{
\dontrun{
# read and into a csv
xlsx_dir <- system.file("xlsx-examples", package = "misc")
read_all_xlsx_then_save_csv(
  path_to_xlsx = xlsx_dir
)
}

}
