% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_all_xlsx_then_save_csv.R
\name{read_all_xlsx_then_save_csv}
\alias{read_all_xlsx_then_save_csv}
\title{Read all sheets from all excel files and save into csv files}
\usage{
read_all_xlsx_then_save_csv(path_to_xlsx)
}
\arguments{
\item{path_to_xlsx}{a character vector with the path to excel file}
}
\description{
Read all sheets from all excel files and save into csv files
}
\examples{
\dontrun{
# read and into a csv
xlsx_dir <- system.file("xlsx-examples", package = "misc")
read_all_xlsx_then_save_csv(
  path_to_xlsx = xlsx_dir
)
}

}
