% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sheet_then_save_csv.R
\name{read_sheet_then_save_csv}
\alias{read_sheet_then_save_csv}
\title{Read a excel sheet and save into a csv file}
\usage{
read_sheet_then_save_csv(excel_sheet, path_to_xlsx, dir_to_save = NULL)
}
\arguments{
\item{excel_sheet}{a character vector with the name of the excel sheet}

\item{path_to_xlsx}{a character vector with the path of the excel file}

\item{dir_to_save}{a character vector with the path to save the csv file. Default is NULL and save the csv in the "data/temp" if it exists.}
}
\description{
Read a excel sheet and save into a csv file
}
\examples{
\dontrun{
# read and into a csv
misc::create_dirs("ma-box")
xlsx_file <- system.file("xlsx-examples", "mtcars_workbook_001.xlsx", package = "misc")
read_sheet_then_save_csv(
  excel_sheet = "mtcars_sheet_001",
  path_to_xlsx = xlsx_file,
  dir_to_save = "ma-box"
)
}
}
