% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_temp_data.R
\name{save_temp_data}
\alias{save_temp_data}
\title{Save objects as a RDS file}
\usage{
save_temp_data(object, dir_to_save = NULL)
}
\arguments{
\item{object}{R object}

\item{dir_to_save}{a character vector with the directory name. Default is NULL and save object in the "data/temp" if it exists.}
}
\description{
Save objects as a RDS file
}
\examples{
\dontrun{
# create and save a R object
awesome <- "not to much!"
misc::create_dirs("ma-box")
save_temp_data(object = awesome, dir_to_save = "ma-box")
# using default directories from `misc::create_dirs()`
create_dirs()
so_good <- "Yep!"
save_temp_data(object = so_good)
}
}
