% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_MultiTest.R
\name{stats_MultiTest}
\alias{stats_MultiTest}
\alias{stats_MultiTest.formula}
\alias{stats_MultiTest.lm}
\alias{stats_MultiTest.glm}
\alias{stats_MultiTest.polr}
\title{Multiple elements Analysis for data.frame}
\usage{
stats_MultiTest(x, ...)

\method{stats_MultiTest}{formula}(formu, dat, ...)

\method{stats_MultiTest}{lm}(obj)

\method{stats_MultiTest}{glm}(obj)

\method{stats_MultiTest}{polr}(obj)
}
\arguments{
\item{...}{other parameters for \code{lm}, \code{glm} or \code{MASS::polr}.}

\item{dat}{data.frame}

\item{obj}{\code{lm}, \code{glm} or \code{polr} object.}

\item{formula}{formula, like \code{glm}}
}
\value{
a \code{stat_MultiTest} object.
}
\description{
Multiple elements Analysis for data.frame
}
\examples{
stats_MultiTest(mpg~., mtcars)
}
