% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tad_view.R
\name{tad_view}
\alias{tad_view}
\title{Alternative data.frame viewer using tad}
\usage{
tad_view(data)
}
\arguments{
\item{data}{a data.frame/tibble data format.}
}
\value{
None
}
\description{
\code{tad_view()} is an alternative to \code{View()} function when not using
RStudio. Please, make sure you have
\href{https://github.com/antonycourtney/tad}{tad} installed in your
system.
}
\examples{
\donttest{
if (interactive()) {
  library(misc)
  mtcars \%>\%
    tad_view()
}
}
}
\seealso{
Other data-viewers:
\code{\link[=view_excel]{view_excel()}},
\code{\link[=view_in]{view_in()}},
\code{\link[=view_vd]{view_vd()}},
\code{\link[=view_vd_nonint]{view_vd_nonint()}}
}
\concept{data-viewers}
