% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tad_view.R
\name{tad_view}
\alias{tad_view}
\title{Alternative \code{View()} function}
\usage{
tad_view(data)
}
\arguments{
\item{data}{a data.frame/tibble data format.}
}
\value{
None
}
\description{
\code{tad_view()} is an alternative to \code{View()} function when not using
RStudio. Please, make sure you have
\href{https://github.com/antonycourtney/tad}{tad} installed in your
system.
}
\examples{
\dontrun{
library(tidyverse)
mtcars \%>\%
  tad_view()
}
}
