% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_fig.R
\name{trim_fig}
\alias{trim_fig}
\title{Remove white spaces around figures}
\usage{
trim_fig(figure_path)
}
\arguments{
\item{figure_path}{a character vector with path of the figure}
}
\description{
Remove white spaces around figures
}
\examples{
\dontrun{
library(misc)
ipak(c("ggplot2", "dplyr"))
create_dirs()
p <- mtcars \%>\%
  ggplot() +
  aes(x = mpg, y = cyl) +
  geom_point()
save_plot(p)
trim_fig("output/figures/p.png")
}
}
