% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_in.R
\name{view_excel}
\alias{view_excel}
\title{View data frame in Excel or other spreadsheet viewer}
\usage{
view_excel(data, viewer = c("excel", "libreoffice", "gnumeric", "tad"))
}
\arguments{
\item{data}{A data frame to view}

\item{viewer}{The spreadsheet viewer to use. One of \code{"excel"} (default), \code{"libreoffice"},
\code{"gnumeric"}, or \code{"tad"}.}
}
\value{
Returns nothing
}
\description{
Opens a data frame in Microsoft Excel or another spreadsheet viewer. Also copies the data
to the system clipboard.
}
\seealso{
Other data-viewers:
\code{\link[=serve_plots]{serve_plots()}},
\code{\link[=serve_plots_url]{serve_plots_url()}},
\code{\link[=tad_view]{tad_view()}},
\code{\link[=view_in]{view_in()}},
\code{\link[=view_qgis]{view_qgis()}},
\code{\link[=view_vd]{view_vd()}},
\code{\link[=view_vd_nonint]{view_vd_nonint()}}
}
\concept{data-viewers}
