% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_in.R
\name{view_excel}
\alias{view_excel}
\title{View data frame in Excel or other spreadsheet viewer}
\usage{
view_excel(data, viewer = c("libreoffice", "gnumeric", "tad", "excel"))
}
\arguments{
\item{data}{A data frame to view}

\item{viewer}{The spreadsheet viewer to use. One of "libreoffice", "gnumeric", "tad", or "excel" (default)}
}
\value{
Returns nothing
}
\description{
Opens a data frame in Microsoft Excel or another spreadsheet viewer. Also copies the data
to the system clipboard.
}
