% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_in.R
\name{view_in}
\alias{view_in}
\title{Alternative data.frame viewer}
\usage{
view_in(data, viewer = c("libreoffice", "gnumeric", "tad"))
}
\arguments{
\item{data}{a data.frame/tibble data format.}

\item{viewer}{character app to open the csv file.}
}
\value{
None
}
\description{
\code{view_in()} is an alternative to \code{View()} function when not using
RStudio. To date, it works with gnumeric, libreoffice and tad.
}
\examples{
\donttest{
if (interactive()) {
  library(misc)
  mtcars \%>\%
    view_in()
}
}
}
\seealso{
Other data-viewers:
\code{\link[=tad_view]{tad_view()}},
\code{\link[=view_excel]{view_excel()}},
\code{\link[=view_qgis]{view_qgis()}},
\code{\link[=view_vd]{view_vd()}},
\code{\link[=view_vd_nonint]{view_vd_nonint()}}
}
\concept{data-viewers}
